refactor: move api/system to api/public
+ replaced api/system with api/public router
+ changed prefix accordingly
+ make relevant changes in frontend
+ make relevant changes in docker-compose.*.yml
+ do not forget to change your healthcheck url accordingly if using provided docker-compose with healthcheck for api
+ that is not breaking change but healthcheck will swear

Related to #28
