+++ backend/app/api/__init__.py
@@ -2,5 +2,5 @@ from .auth import router as auth_router
 from .user import router as user_router
 from .card import router as card_router
 from .password_recovery import router as password_recovery_router
-from .system import router as system_router
+from .public import router as public_router
 from .admin import router as admin_router
