+++ frontend/src/app/entities/public/public-api.service.ts
@@ -1,20 +1,20 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
-import { IVersion } from './system-interface';
+import { IVersion } from './public-interface';
 import { environment } from '@env/environment';
 
 @Injectable({
   providedIn: 'root',
 })
-export class SystemApiService {
+export class PublicApiService {
   constructor(private readonly httpClient: HttpClient) {}
 
   version(): Observable<IVersion> {
-    return this.httpClient.get<IVersion>(`${environment.api}/system/version`);
+    return this.httpClient.get<IVersion>(`${environment.api}/public/version`);
   }
 
   health(): Observable<any> {
-    return this.httpClient.get(`${environment.api}/system/health`);
+    return this.httpClient.get(`${environment.api}/public/health`);
   }
 }
