+++ frontend/src/app/entities/public/public-interface.ts
