+++ frontend/src/app/features/admin/admin-status/admin-status.component.ts
@@ -7,7 +7,7 @@ import { TranslateModule, TranslateService } from '@ngx-translate/core';
 import { map, retry, catchError, of, share } from 'rxjs';
 import { SnackService } from 'src/app/core/services/snack.service';
 import { AdminApiService } from 'src/app/entities/admin/admin-api.service';
-import { SystemApiService } from 'src/app/entities/system/system-api.service';
+import { PublicApiService } from 'src/app/entities/public/public-api.service';
 
 @Component({
   selector: 'app-admin-status',
@@ -26,9 +26,9 @@ export class AdminStatusComponent {
   private readonly translateService = inject(TranslateService);
   private readonly adminApiService = inject(AdminApiService);
   private readonly snackService = inject(SnackService);
-  private readonly systemApiService = inject(SystemApiService);
+  private readonly publicApiService = inject(PublicApiService);
 
-  public readonly health$ = this.systemApiService.health().pipe(
+  public readonly health$ = this.publicApiService.health().pipe(
     map(() => true),
     retry({ count: 1, delay: 1000 }),
     catchError((error) => {
